Deprecate AMI image formats
We have a lot of special cases all around our code for AMI (et al)
images, in the boot, snapshot, and other areas. It was recently
discussed (after the major CVE which was complicated by AMI support)
that we should drop these special cases and support for AMI with it.
This starts that timer.

Change-Id: I90efaaa35225ce82bada0b427e09c8289adaae1e