api: Handle empty imageRef alongside null for local BDM check
We don't want users creating BDMs with source_type=image,
destination_type=local and boot_index=0, since that redefines our root
disk definition which should be controlled by the flavor's 'disk'
property. We did have a check for this, but it explicitly checked for
the 'imageRef' being set to null - which isn't permitted by the schemas
-instead of the empty string that is expected when creating a BFV
instance. Correct this check.

Change-Id: Id3442c3d315e6bb63e6c3675f4ab104885b3884f
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes-Bug: #2077980