api: Add response body schemas for hypervisors APIs (1/3)
We split this one up due to its size, which itself is mainly due to the
amount of aliasing that went on in early versions as well as the amount
of changes that have been made over the years.

This change focuses on the index view and related non-detailed views.

Change-Id: I49828e17f9140da16c02fc44fe446c4e489b4d77
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>