docs: Show the recommended way to PXE boot an instance
Someone asked about this in the #openstack-nova channel today and I
found some artifacts: an abandoned patch [1] and spec [2] where the
conclusion of the discussion was a recommendation to use iPXE and
rescue to PXE boot an instance.

This adds a small high level doc about how to do it to help others who
may want to do the same thing.

[1] https://review.opendev.org/c/openstack/nova/+/434549
[2] https://review.opendev.org/c/openstack/nova-specs/+/435052

Change-Id: I78f097b9e280fc2bcc14655042b844d3d5977b07