Refactor obj_make_compatible to reduce complexity
Our tox.ini sets the maximum complexity of a function to be 40. Adding
anything new to the ImageMetaProps.obj_make_compatible push it over the
limit. So this patch factors out common new field value checking logic
to a helper function to reduce the complexity before adding new supported
values.

As there was a small amount of this check in other ovos this patch does
the refactor across all nova ovos.

Change-Id: I828d91196ea06f4dc2823218f036960123446255