Add service version for Dalmatian
We agreed by I2dd906f34118da02783bb7755e0d6c2a2b88eb5d  on the support
envelope.
Pre-RC1, we need to add a service version in the object.
Post-RC1, depending on whether it's SLURP or not SLURP, we need to bump
the minimum version or not.

This patch only focuses on pre-RC1 stage.
Given Epoxy won't be skippable, we won't need a post-RC1 patch for updating the min
that will continue to support Caracal.

HTH.

Change-Id: I9737a6ea02203605c1d36204b74fe3575764841e