Respect supplied arguments in novncproxy_base_url
In case an operator want to have NoVNC under a certain subpath instead
of the subdomain or custom port, they would need to instruct NoVNC
to establish WebSocket connection with a correct URI, which is passed
through `path` argument.

The implements parsing of novncproxy_base_url and appending token
to the request rather then re-writing request completely.

Implements: blueprint novnc-base-url-respect-extra-params
Change-Id: Ie900f4963a998222942c3d54a757ef1e625f7bf9