Handle iso+gpt detections
Some bootable CD images have an MBR or GPT boot record in the system
area of the ISO format. That causes us to detect both ISO and GPT
formats, which are currently rejected (intentionally). Since that
special case is legitimate and we need to treat those as ISO files,
this adds handling to make that determination. It does so by handling
the detection pipeline within nova so we have access to the multiple
matching inspectors (similar to what glance does).

Depends-On: https://review.opendev.org/c/openstack/requirements/+/934176
Change-Id: I01e4f1bd74c9535f1e588159fd5e91c9b8bc60d4