Fix detaching devices by alias with mdevs
This change fixes the get_device_by_alias function
to properly handle devices that do not have an
alias in the domain xml.

By definition if a device does not have an alias
it can't match a given alias, so this change adds
a hasattr check to ensure we do not attempt to check
the alias if it's not defined.

This eliminates this entire class of bugs for all
devices instead of just fixing it for a specific class.

Related-Bug: #1942345
Closes-Bug: #2074219
Change-Id: If417a43ea252647618e50391b63333f6b68bdfec
(cherry picked from commit 9db3eba8133f4f643bc4a748460a97434fa91f78)