Fix case sensitive comparison
As of now, if flavor has 'hw:virtio_packed_ring' extra-spec set
to True and image has 'hw_virtio_packed_ring' set to 'true', if
fails the comparison and throws exception.

This change makes the comparison case-insensitive by adding
another condition for image_value to change it from str
to bool if its not None.

Closes-Bug: #2085124
Change-Id: Ie1af94b216f10cad592f38037875744927c4f18a