Stop asking bug report to upstream
This was useful in the past when nova was actively developed and more
people started testing it. However the situation is different recently
and we've seen tons of actual production use case, and we have seen
that the issues are mostly specific to wrong deployment/usage rather
than something caused by actual code bug in nova.

Change-Id: If212c03e9c33bb0644a72cedee61b33598097544