api: Add response body schemas for hypervisors APIs (2/3)
We split this one up due to its size, which itself is mainly due to the
amount of aliasing that went on in early versions as well as the amount
of changes that have been made over the years.

This focuses on the show view as well as the closely related detailed
view.

Change-Id: I06b1a8f0a30e9260f118921a13e85c4c534d9c8f