vmware: Switch mks console to `webmks`
Since vSphere 6, VMware supports `webmks` as `ticketType` in addition to
`mks`. `webmks` provides the remote console of a VM via WebSocket
endpoint and is supposed to be used with their SDK. It does fall back to
normal VNC which can be used with e.g. `noVNC` though.

With this change, we can follow-up by removing our custom mks proxy
code, as we plan to expose `webmks` via Nginx directly.

Change-Id: I99294d7f6bf0539d20887322eeadb2580297fc8a