Refactor test_server_shares: Mock in Base Class and trait verification
Mock the connect and disconnect methods in the base class to centralize
and streamline calls.

Additionally, create a reusable function to verify traits

Manila is the OpenStack Shared Filesystems service.
These series of patches implement changes required in Nova to allow
the shares provided by Manila to be associated with and attached to
instances using virtiofs.

Implements: blueprint libvirt-virtiofs-attach-manila-shares
Change-Id: If8b2263c86f4eb580426b1a7b3de28c85cf5af1a