Update contributor process documentation
This patch updates the contributor/process.rst file to remove outdated
process descriptions, such as runways and gerrit review labels.

The patch also adds the description of the current tracking etherpad
that the team uses to track priorities and ongoing work items. It also
adds the link to the process.rst doc to README.rst for improved
visibility.

Closes-Bug: #2089325

Change-Id: I1713f84fc521759333d1d07fc12d0fc23a5cb807
Signed-off-by: Ildiko Vancsa <ildiko.vancsa@gmail.com>