Support in-cluster (anti-)affinity for VMware
We expose a capability "resource_scheduling" which enables the
`ComputeManager` to decide whether to allow multiple VMs on the same
host in the post-scheduling check.

We also update the DRS rules when spawning/resizing/migrating a VM.

"soft-affinity" is excluded from DRS rule creation and thus only works
on cluster-level not on HV-level. The reason is, that DRS doesn't know
about "soft" and will require the VMs to run on the same ESXi. Having
the VMs run inside the same rack should be good-enough though.

For managing DRS settings, we introduce a new module `cluster_util` to
the `vmwareapi` driver.

Change-Id: Iba9706b7353575fd3f5df5e903e3f254c5e49d73
Co-Authored-By: Johannes Kulik <johannes.kulik@sap.com>
Co-Authored-By: Fabian Wiesel <fabian.wiesel@sap.com>
Incorporated-Change-Id: I6d44910c420de4c76b6112904ccfebe3ec923098