Update InstanceNUMACell version after data migration
The data migration of InstanceNUMACell 1.4 to 1.5 only moved the data to
the new pcpuset field but does not update the ovo version string of the
object in the DB. This resulted in an 1.6 data with an 1.4 version
string in the DB which subsequently causes that an old compute running
1.4 ovo version will think it got an old ovo even though the data is
already in the new format. This leads to instance lifecycle errors and
if the nova-compute saves the instance then it also leads to permanent
data loss.

So this change modified the data migration to also update the ovo version
in the DB so that the version string in the DB matches the schema the
data uses in the DB.

Related-Bug: #2097359
Change-Id: I9a99f10526f8e466ac04b035121b24be70a23dae
(cherry picked from commit 643a6a8a57945ceb1bf3c5971be2aeb3efd4fa72)