Add Project Manager role context in unit tests
As part of the this BP[1], we are going to add the
Project Manager role as default in API policies.

This commit adds the project manager role context
in the policies unit tests so that when we change
the policy default to project manager then we can
know how the changes/permission will looks like.

NOTE: This does not change any policy default, only
adding manager context in existing unit tests.

Partial implement blueprint policy-manager-role-default

[1] https://blueprints.launchpad.net/nova/+spec/policy-manager-role-default

Change-Id: Ib1622e7593108aa4fac65fcc34cb616502abed94
Signed-off-by: Ghanshyam Mann <gmaan@ghanshyammann.com>