Update conductor and filters allowing migration with SR-IOV devices
This patch enable live migration for SR-IOV PCI devices that are
live-migratable.
A ``live_migratable`` tag is added to the device specification
in [pci]dev_spec config.
A new filter is added to stats.py in the filter_pools() function to
handle PCI requests for live_migratable devices.

The target goal of these series of patch is to enable VFIO devices
migration with kernel variant drivers.

Partially-Implements: blueprint migrate-vfio-devices-using-kernel-variant-drivers
Change-Id: Ifbd19947ba992589a00fdf7527380450aa328269