Add fill_metadata() to InstanceList
This adds a non-remotable method to InstanceList which will batch-
fill system_metadata for all the instances in the list that are
missing it in as efficient of a manner as possible. This does not
require an object bump because no remotable methods or fields are
changed.

Related to blueprint image-metadata-props-weigher

Change-Id: Icc47de2b677b3d212a7f6faa61a85ea9bff9f412