Fix issue with "disable db-only delete" when using --force
With "disable db only delete" enabled, when using force delete, a
ComputeServiceUnavailable exception is raised but not caught. This
change fixes this, catching the exception and reporting the same error
message to the user as if they used delete without force.

fixup-for: Change behavior for deleting servers when nova-compute is down
Change-Id: I3a5e8d28d519da3f6dfe8ed194c28e346499e28c