vmware: Support percentage-based HA policy
Until this commit, we supported reporting reserved resources to
Placement based on the admission control policy setting for failover
hosts. Since failover hosts are required for big VMs and we were
previously running in mixed clusters with small VMs, we had that
admission control policy set everywhere.

In the future, we want support running general purpose clusters without
big VMs. For these clusters, failover hosts are not a requirement, so we
implement the percentage based admission control policy.

Since we still have code that supports default reservations (for the
ESXi itself) and DRS host group based reservations, we have to integrate
these with the reservations of the HA policy. Since HA is cluster based
and does not care for host groups, we have to add the reservations to
every host in the cluster. Hence, we update our default/DRS group based
reservations dictionary to include the percentage.

One problem with this approach is, that using non-percentage based
default/host group based reservations instead of percentage-based ones
will ignore the HA policies. This is something we silently ignore
currently, because we do not use host group based policies at all and
reserve by default 5 % of the ESXi for itself - not a specific amount.

Change-Id: I5dee0767b5e5b9306cbf2da1238f22a0c8181cc6