unified limits: discover service ID and region ID
In oslo.limit 2.6.0 service endpoint discovery was added, provided by
three new config options:

  [oslo_limit]
  endpoint_service_type = ...
  endpoint_service_name = ...
  endpoint_region_name = ...

We can use the same config options if they are present to lookup the
service ID and region ID we need when calling the
GET /registered_limits API as part of the resource limit enforcement
strategy. This way, the user will not have to configure endpoint_id.

This will look for [oslo.limit]endpoint_id first and if it is not set,
it will do the discovery.

Closes-Bug: #1931875

Change-Id: Ida14303115e00a1460e6bef4b6d25fc68f343a4e